Add zlib as an explicit dependency
It appears that the inclusion of libxml2 doesn't necessarily mean
that zlib will be included. Since libcomps depends on zlib functions,
the build fails when zlib isn't being brought in for the build.

This explicitly changes it to include zlib and ensure it is linked
into the library at build-time.