libcomps/CMakeLists.txt: up cmake version requirement to 3.0.2+.
We already use features that are only available in cmake 3.0+ (e.g., the
GNUInstallDirs module), so might as well acknowledge that.

Note that this means losing support for ancient distros, e.g., RHEL 6,
which ships cmake 2.8.12.2, but given the fact above that's a moot
point.

The good part is that EPEL 6 ships a rather recent cmake version
(3.6.1), so building the package on RHEL 6 is still very much possible
when using the EPEL 6 repository at build time.