Do not skip type=mandatory in xml output and test it (RhBug:1771224)
The mandatory type is used when no other configuration and no other
type is specified but since this behavior can be overridden with passed
configuration it doesn't make sense to always skip outputting the type
mandatory.
The user can specify for example optional type as default and then
the knowledge that some specific <packagereq> were mandatory would be
lost.

https://bugzilla.redhat.com/show_bug.cgi?id=1771224