Fix non-optimized builds by removing `inline` keyword
The raw (without `static` or `extern`) `inline` keyword defines an
alternative version of a functions that is used when inlining.
When optimizations are turned off (eg when `-DCMAKE_BUILD_TYPE=Debug`)
it wants to use the original version of the function but there is no such
definition -> linking fails.

To fix this and preserve API remove the `inline` keyword and define the
functions normally.